Absorption of an OpenSystem
Absorption spectrum can be calculated for an OpenSystem. Natural lifetime calculation moved to OpenSystem.
