+++ quantarhei/builders/__init__.py
@@ -5,6 +5,7 @@ from .sysmodes import AnharmonicMode
 from .sysmodes import HarmonicMode
 from .molecules import Molecule
 from .aggregates import Aggregate
+from .opensystem import OpenSystem
 from .pdb import PDBFile
 from .aggregate_states import ElectronicState
 from .aggregate_states import VibronicState
