+++ quantarhei/spectroscopy/twodcontainer.py
@@ -9,7 +9,6 @@
 """
 import numbers
 
-#import h5py
 #import matplotlib.pyplot as plt  
 import numpy
